Check if Caddyfile exists.

+++ docker-entrypoint.sh
@@ -2,7 +2,8 @@
 
 set -e
 
-if [ ! -z "$CADDYFILE" ]; then
+if [ ! -z "$CADDYFILE" ] && [ ! -e "$CADDYFILE" ] ; then
+  echo "Copy initial Caddyfile to" "$CADDYFILE"
   cp /etc/Caddyfile "$CADDYFILE"
 fi
